Move Image-related code into its own file
