+++ tools/internal/config/config.go
@@ -16,24 +16,6 @@ type Config struct {
 	Repositories []Repository
 }
 
-// Image should not be instantiated directly. Instead, use NewImage().
-type Image struct {
-	// If true, the Image is not mirrored i.e. it is not added to the
-	// regsync config when the regsync config is generated.
-	DoNotMirror any `json:",omitempty"`
-	// The source image without any tags.
-	SourceImage            string
-	defaultTargetImageName string
-	// Used to specify the desired name of the target image if it differs
-	// from default. This field would be private if it was convenient for
-	// marshalling to JSON/YAML, but it is not. This field should not be
-	// accessed directly - instead, use the TargetImageName() and
-	// SetTargetImageName() methods.
-	SpecifiedTargetImageName string `json:"TargetImageName,omitempty"`
-	// The tags that we want to mirror.
-	Tags []string
-}
-
 type Repository struct {
 	// BaseUrl is used exclusively for building the target image ref
 	// for a given image for a repository. For example, a target
@@ -145,32 +127,6 @@ func (config *Config) ToRegsyncConfig() (regsync.Config, error) {
 	return regsyncYaml, nil
 }
 
-// ToRegsyncImages converts image into one ConfigSync (i.e. an image
-// for regsync to sync) for each tag present in image. repo provides
-// the target repository for each ConfigSync.
-func (image *Image) ToRegsyncImages(repo Repository) ([]regsync.ConfigSync, error) {
-	entries := make([]regsync.ConfigSync, 0, len(image.Tags))
-	for _, tag := range image.Tags {
-		doNotMirror, err := image.DoNotMirrorTag(tag)
-		if err != nil {
-			return nil, fmt.Errorf("failed to determine whether tag %q should be mirrored: %w", tag, err)
-		}
-		if doNotMirror {
-			continue
-		}
-		sourceImage := image.SourceImage + ":" + tag
-		targetImage := repo.BaseUrl + "/" + image.TargetImageName() + ":" + tag
-		entry := regsync.ConfigSync{
-			Source: sourceImage,
-			Target: targetImage,
-			Type:   "image",
-		}
-		entries = append(entries, entry)
-	}
-
-	return entries, nil
-}
-
 func CompareImages(a, b *Image) int {
 	if sourceImageValue := strings.Compare(a.SourceImage, b.SourceImage); sourceImageValue != 0 {
 		return sourceImageValue
@@ -180,85 +136,4 @@ func CompareImages(a, b *Image) int {
 
 func compareRepositories(a, b Repository) int {
 	return strings.Compare(a.BaseUrl, b.BaseUrl)
-}
-
-func NewImage(sourceImage string, tags []string) (*Image, error) {
-	image := &Image{
-		SourceImage: sourceImage,
-		Tags:        tags,
-	}
-	if err := image.setDefaults(); err != nil {
-		return nil, err
-	}
-	return image, nil
-}
-func (image *Image) Sort() {
-	slices.Sort(image.Tags)
-}
-
-func (image *Image) setDefaults() error {
-	parts := strings.Split(image.SourceImage, "/")
-	if len(parts) < 2 {
-		return fmt.Errorf("source image split into %d parts (>=2 parts expected)", len(parts))
-	}
-
-	if parts[0] == "dp.apps.rancher.io" {
-		// AppCo images have only one significant part in their reference.
-		// For example, in dp.apps.rancher.io/containers/openjdk,
-		// dp.apps.rancher.io/containers is the repository and openjdk is
-		// the significant part.
-		imageName := parts[len(parts)-1]
-		image.defaultTargetImageName = "appco-" + imageName
-	} else {
-		repoName := parts[len(parts)-2]
-		imageName := parts[len(parts)-1]
-		image.defaultTargetImageName = "mirrored-" + repoName + "-" + imageName
-	}
-	return nil
-}
-
-func (image *Image) TargetImageName() string {
-	if image.SpecifiedTargetImageName != "" {
-		return image.SpecifiedTargetImageName
-	}
-	return image.defaultTargetImageName
-}
-
-func (image *Image) SetTargetImageName(value string) {
-	if value == image.defaultTargetImageName {
-		image.SpecifiedTargetImageName = ""
-	} else {
-		image.SpecifiedTargetImageName = value
-	}
-}
-
-func (image *Image) CombineSourceImageAndTags() []string {
-	fullImages := make([]string, 0, len(image.Tags))
-	for _, tag := range image.Tags {
-		fullImage := image.SourceImage + ":" + tag
-		fullImages = append(fullImages, fullImage)
-	}
-	return fullImages
-}
-
-func (image *Image) DoNotMirrorTag(tag string) (bool, error) {
-	switch val := image.DoNotMirror.(type) {
-	case nil:
-		return false, nil
-	case bool:
-		return val, nil
-	case []any:
-		for _, valPart := range val {
-			blacklistedTag, ok := valPart.(string)
-			if !ok {
-				return false, fmt.Errorf("failed to cast %v to string", valPart)
-			}
-			if blacklistedTag == tag {
-				return true, nil
-			}
-		}
-		return false, nil
-	default:
-		return false, nil
-	}
 }
