+++ tools/internal/config/image_test.go
